Address review comments

+++ src/eip7594/eip7594.c
@@ -457,25 +457,21 @@ out:
 /**
  * Aggregate columns, compute the sum of interpolation polynomials, and commit to the result.
  *
- * This function computes: [sum_k r^k interpolation_poly_k(s)]
+ * This function computes `RLI = [sum_k r^k interpolation_poly_k(s)]` from the spec.
  *
- * This function first aggregates cells from the same column by scaling them with the corresponding
- * powers of the random challenge. It then computes the sum of the interpolation polynomials from
- * the aggregated columns and finally commits to the aggregated interpolation polynomial.
- *
- * @param[out]  commitment_out              Commitment to the aggregated interpolation poly
- * @param[in]   r_powers                    Precomputed powers of the random challenge
- * @param[in]   cell_indices                Indices of the cells
- * @param[in]   cells                       Array of cells
- * @param[in]   num_cells                   Number of cells
- * @param[in]   s                           The trusted setup
+ * @param[out]  commitment_out      Commitment to the aggregated interpolation poly
+ * @param[in]   r_powers            Precomputed powers of the random challenge
+ * @param[in]   cell_indices        Indices of the cells
+ * @param[in]   cells               Array of cells
+ * @param[in]   num_cells           Number of cells
+ * @param[in]   s                   The trusted setup
  */
 static C_KZG_RET compute_commitment_to_aggregated_interpolation_poly(
     g1_t *commitment_out,
     const fr_t *r_powers,
     const uint64_t *cell_indices,
     const Cell *cells,
-    size_t num_cells,
+    uint64_t num_cells,
     const KZGSettings *s
 ) {
     C_KZG_RET ret;
@@ -484,7 +480,14 @@ static C_KZG_RET compute_commitment_to_aggregated_interpolation_poly(
     fr_t *aggregated_interpolation_poly = NULL;
     bool *is_cell_used = NULL;
 
-    // Allocate memory
+    /*
+     * This function first aggregates cells from the same column by scaling them with the
+     * corresponding powers of the random challenge. It then computes interpolation polynomials
+     * using the aggregated cells. It finally sums up the interpolation polynomials and commits to
+     * the aggregated interpolation polynomial.
+     */
+
+    /* Allocate memory */
     ret = new_fr_array(&aggregated_column_cells, FIELD_ELEMENTS_PER_EXT_BLOB);
     if (ret != C_KZG_OK) goto out;
     ret = new_fr_array(&column_interpolation_poly, FIELD_ELEMENTS_PER_CELL);
@@ -494,7 +497,7 @@ static C_KZG_RET compute_commitment_to_aggregated_interpolation_poly(
     ret = new_bool_array(&is_cell_used, FIELD_ELEMENTS_PER_EXT_BLOB);
     if (ret != C_KZG_OK) goto out;
 
-    // Initialize aggregated columns and usage flags
+    /* Start with zeroed out columns */
     for (size_t i = 0; i < CELLS_PER_EXT_BLOB; i++) {
         for (size_t j = 0; j < FIELD_ELEMENTS_PER_CELL; j++) {
             size_t index = i * FIELD_ELEMENTS_PER_CELL + j;
@@ -503,8 +506,8 @@ static C_KZG_RET compute_commitment_to_aggregated_interpolation_poly(
         }
     }
 
-    // Aggregate the cells by scaling and summing
-    for (size_t i = 0; i < num_cells; i++) {
+    /* Scale each cell's data points */
+    for (uint64_t i = 0; i < num_cells; i++) {
         for (size_t j = 0; j < FIELD_ELEMENTS_PER_CELL; j++) {
             fr_t cell_fr, scaled_fr;
             size_t offset = j * BYTES_PER_FIELD_ELEMENT;
@@ -521,32 +524,45 @@ static C_KZG_RET compute_commitment_to_aggregated_interpolation_poly(
         }
     }
 
-    // Initialize aggregated interpolation polynomial
+    /* Start with a zeroed out poly */
     for (size_t i = 0; i < FIELD_ELEMENTS_PER_CELL; i++) {
         aggregated_interpolation_poly[i] = FR_ZERO;
     }
 
-    // Sum the interpolation polynomials
+    /* Interpolate each column */
     for (size_t i = 0; i < CELLS_PER_EXT_BLOB; i++) {
+        /* Offset to the first cell for this column */
         size_t index = i * FIELD_ELEMENTS_PER_CELL;
 
+        /* We only care about initialized cells */
         if (!is_cell_used[index]) continue;
 
+        /* We don't need to copy this because it's not used again */
         ret = bit_reversal_permutation(
             &aggregated_column_cells[index], sizeof(fr_t), FIELD_ELEMENTS_PER_CELL
         );
         if (ret != C_KZG_OK) goto out;
 
+        /*
+         * Get interpolation polynomial for this column. To do so we first do an IDFT over the roots
+         * of unity and then we scale by the coset factor.  We can't do an IDFT directly over the
+         * coset because it's not a subgroup.
+         */
         ret = fr_ifft(
             column_interpolation_poly, &aggregated_column_cells[index], FIELD_ELEMENTS_PER_CELL, s
         );
         if (ret != C_KZG_OK) goto out;
 
+        /*
+         * To unscale, divide by the coset. It's faster to multiply with the inverse. We can skip
+         * the first iteration because its dividing by one.
+         */
         uint64_t pos = reverse_bits_limited(CELLS_PER_EXT_BLOB, i);
         fr_t inv_coset_factor;
         blst_fr_eucl_inverse(&inv_coset_factor, &s->roots_of_unity[pos]);
         shift_poly(column_interpolation_poly, FIELD_ELEMENTS_PER_CELL, &inv_coset_factor);
 
+        /* Update the aggregated poly */
         for (size_t k = 0; k < FIELD_ELEMENTS_PER_CELL; k++) {
             blst_fr_add(
                 &aggregated_interpolation_poly[k],
@@ -556,7 +572,7 @@ static C_KZG_RET compute_commitment_to_aggregated_interpolation_poly(
         }
     }
 
-    // Commit to the final aggregated interpolation polynomial
+    /* Commit to the final aggregated interpolation polynomial */
     ret = g1_lincomb_fast(
         commitment_out,
         s->g1_values_monomial,
@@ -722,10 +738,10 @@ C_KZG_RET verify_cell_kzg_proof_batch(
     if (ret != C_KZG_OK) goto out;
 
     ////////////////////////////////////////////////////////////////////////////////////////////////
-    // Commmit to aggregated interpolation polynomial: [sum_k r^k interpolation_poly_k(s)]
+    // Commmit to aggregated interpolation polynomial
     ////////////////////////////////////////////////////////////////////////////////////////////////
 
-    /* Aggregated cells from same columns, sum interpolation polynomials, and commit */
+    /* Aggregate cells from same columns, sum interpolation polynomials, and commit */
     ret = compute_commitment_to_aggregated_interpolation_poly(
         &interpolation_poly_commit, r_powers, cell_indices, cells, num_cells, s
     );
